Overhaul to support new tagged generics.
The Typeclasses that are used to power the Waargonaut generics have been updated to have a tag param
for more granular control, so these instances for Servant needed an overhaul.

Waargonaut generic module also reexports tagged and generics-sop things that we needed so we could
drop the explicit dependencies.

THERE ARE TESTS IN HERE NOW!!! OMG. I know, right?!